Implemented client-side rendering.
This is a basic first step, to get the functionality out the door.
Later we should implement more robust path-building of the views via a separate store
for the client-side views.

On the server side, this is currently handled by the filesystem store.

I could imagine a "arraycache" store for the client-side, that is able to build the
required view path from the options. That way, we don't have to separate the logic
in the view renderer between server and client.