Made Tower respect database configuration.
Had to change MongoDB to Mongodb everywhere to match Tower.Support.String.camelize conventions so that the adapter could be found programmatically.

Tower will loop through each database in config/databases and send each the configuration for the current environment (Tower.env). Tower.Model.default('store') will be set based on the following rules:

1. If there's no database configuration found at all Tower.Store.Memory will be used
2. If a database configuration is found, the first adapter listed will be used as the default
3. If a database configuration is found and any of them are configured with `default: true` then the last adapter with that configuration will be used.

Because keys are guaranteed to be in the order they're written in an object it's probably best to always set `default: true` on the appropriate adapter (if you're using more than one).